+++ docs/CHANGELOG.md
@@ -9,7 +9,9 @@ For git tags, official releases are marked like "v1.0.0", while unofficial ones 
 
 * ``langFromDict`` and ``targetLangFromDict`` are now officially coming to the kernel utility functions.
 * When creating or editing a new text, text and title boxes change language accordign to the text's language ([#111](https://github.com/HugoFara/lwt/issues/111)).
-* New classe: ``Text`` (in ``inc/classes/Text.php``).
+* Refactored texts with OOP:
+  * New class: ``Text`` (in ``inc/classes/Text.php``).
+  * New function: ``edit_text_form`` (in ``edit_texts.php``) that takes a Text object as input.
 
 ### Fixed